Added python scripts to rebuild builtin.abc (aka "global.abc") and toplevel.abc (aka "shell.abc"), along with the relevant .cpp and .h files. Added the .abc files to the source repository as well so that they will stay in sync. (The weird dual-name issue is a holdover from living in the Adobe Flash build environment, but it's well past time to eliminate the weirdness... I have decided to standardize on the names that sync with the existing source files.)
Usage:

    cd core
    python builtin.py --> regenerates builtin.abc, .cpp, .h

    cd shell
    python toplevel.py --> regenerates toplevel.abc, .cpp, .h

Also updated runtests.pl to refer to the builtin.abc file as builtin.abc rather than global.abc... and also to no longer barf when recompiling files with $ in their name.

Finally, updated the Mac Xcode projects to stop putting the build results in the weird ../../../../platforms/flash/mac result... again, this is a holdover from living in the Adobe Flash build environment, but doesn't make sense for Tamarin.