Initial port of Traits caching from tamarin-tracing. (r=edwsmith)
Traits include information about method bindings, slot type-and-offset, and interface implementation. Large apps (eg written in Adobe Flex) can instantiate thousands of Traits at runtime; for such an app you might have 3000+ Traits, which amount to ~1MB of memory, of which the binding (etc) information above is about 65%.

Instead of generating all of the above information at AbcParse time, we defer it and keep a cache of said information -- more extensive profiling is necessary, but initial examination shows that a cache of the 32 most-recently-used Traits bindings provides a nearly equivalent performance while saving ~99% of the space devoted to this -- ie, over half a megabyte for cases like the above. ("Nearly equivalent" means within 1-2% on typical benchmarks; more optimization remains to be done so it seems likely that this penalty can be reduced even further.)

Currently the cache uses a random-eviction policy when the cache is full; more sophisticated policies have been experimented with (eg true LRU, bitfield-aging) but the tracking overhead hasn't proved to be worthwhile. Other policies might be worth investigating.

Note that you can set the cache size to "unlimited" (which is actually the default), in which case memory usage and performance is essentially equivalent to the old code, and this is the default behavior (you can call AvmCore::setCacheSizes at any time to readjust the cache size).

Currently the feature is controlled by AVMPLUS_TRAITS_CACHE in avmbuild.h; it's anticipated this flag will remain in place for a short time, until it's shown that caching can be used without penalty in all contexts.

As a side-effect of the changes above, lots of the API to Traits was reworked (eg formerly public members are now accessed via function call). There are still public members in Traits but eventually they should all be insulated by (at least) inline function accessors. There also things that you can't get directly from a Traits, but now must get from the (cached) TraitsBindings structure.

The layout in memory of a ScriptObject shifted; formerly the layout was

    BaseClass [,slots] [,hashtable] [,slots]

where the hashtable for dynamic properties (if any) appeared after the slots of the first class that was dynamic, followed by any slots of subsequent inheritors; this layout made dynamic rebuilding of slot offsets trickier than it needed to be, so now the layout is

    BaseClass [,slots] [,hashtable]

ie the hashtable (if any) is always put at the end of the object, with all slots contiguous in the middle.

Since Traits::destroyInstance needs to know slot types in order to properly destroy slots containing RCObjects, added a new bitfield to Traits to mark which slots need this attention so that we don't have to potentially re-create TraitsBindings during object destruction. (Bitfield usage was encapsulated in the new FixedBitSet class, which is a tiny nongrowable bitfield utility that takes no more space than a pointer for small objects, thus is preferable in this case to the existing BitSet class).

Brought over the concept of "BuiltinType" from TT -- the various well-known builtin classes now have an enumeration for their type, meaning you don't have to compare them against (say) core->object_itraits to know it's an Object, and thus don't have to pass AvmCore everywhere you need to do this check. It also means that you can now use switch statements in some places you formerly had to use repeated if-else, and since there are <32 of them, you can also use bitmasking techniques to check for set membership, eg:

   const uint32_t NUMERIC_TYPE_MASK = (1<<BUILTIN_int) | (1<<BUILTIN_uint) | (1<<BUILTIN_number);
   inline bool isNumeric() const { return ((1<<builtinType) & NUMERIC_TYPE_MASK) != 0; }

There is also a disabled-by-default chunk of code controlled by AVMPLUS_TRAITS_MEMTRACK used for advanced memory profiling of the caching scheme; this will probably go away in a few weeks, but for now is still being used for subsequent memory investigations — leave it disabled but please don’t remove it yet.

You can experiment with different cache sizes by calling AvmCore::setCacheSizes, or on the command line of avmshell via:

   -cache_bindings N size of bindings cache (0 = unlimited)
   -cache_metadata N size of metadata cache (0 = unlimited)

If omitted, -cache_bindings defaults to 0 (unlimited) and -cache_metadata to 1.

Other changes that are currently included in this patch (some related, some
not):

-- added an "empty" GCWeakRef singleton to GC (GC::emptyWeakRef) -- this is a weak-ref that always points to null, but is not itself null. This allows you to pre-initialize a GCWeakRef* field to a safe "null" reference that isn't itself null, so you can skip a pointless null check on subsquent use.

-- Also moved slot-and-method Metadata from Traits into a separate cache. You can set this cache to any size you like, but it defaults to a size of 1 since Metadata is rarely examined, and AFAIK never in a time-critical way. (In fact we probably could skip the caching entirely and simply regenerate it on demand, but the caching machinery put in place for bindings makes it trivial and cheap to cache.)

-- slightly sped up getSlotAtom/setSlotAtom for the AVMPLUS_TRAITS_CACHE case: we now pack offset (29 bits) and slot-storage-type (3 bits) into a single 32-bit word, so we can get by with a single load and shift to decide what to do. the storage-type for Atom (the most common case) is 0 so we don't even need an explicit test for it anymore, at least on x86. (Interestingly, repeated if-else testing is faster than switch in this case, at least on GCC 4.0, because not all slot types are equally likely to occur)

-- AbcEnv and DomainEnv were reworked to hide the public members in favor of explicit API calls. We're trying to gradually hide the internals of Tamarin to client apps; this is a baby step in that direction.

-- various changes of "naked" C int types to C99 int types (eg unsigned int -> uint32_t) where it seemed appropriate.

-- input-only Multiname arguments were changed (generally) from Multiname* to const Multiname&. Not complete yet but will be fully transitioned someday.

-- changed Binding and CodeContext to be anonymous pointer types, rather than intptr variations, so they can no longer be accidentally intermixed with Atoms (or ints) as they were before. Provided clean accessors for binding-kind, etc,
and modified code as necessary. (This should be done for Atom too, someday, as was done in TT, but the code churn is high)

-- added "exit(-1);" to the Mac version of GCDebugMsg and AvmDebugMsg after the DebugStr call to ensure that an assertion failure would terminate the app, to match behavior on Windows.

-- added AVMPLUS_SSE2_ALWAYS flag and modified a few bits of code to use it (eg doubleToAtom) -- previously we correctly marked all AMD64 code as being able to assume SSE2, but all MacTel machines (even non-AMD64) can also assume SSE2.

-- added AvmCore::skipU30() when you want to call readU30 and ignore the result. This is mostly a wash for a single call, but a modest performance win if you are calling it repeatedly.

-- reworked a bit of code in MultinameHashtable to improve the speed of the MH_CACHE1 version by inlining some of the checks. then, ended up disabling MH_CACHE1 entirely as none of our current performance suites show a meaningful performance difference, even with the optimization above. (Need to consult with the original author to find the use-cases it improved.)

-- minor cleanup of the 64-bit Mac XCode projects to get them building and (mostly) running in Interpreter-only mode.