Remove lptr longjmp hack, exception info stored in core->exceptionAddr (bug 464643 r=lhansen+, stejohns+)
lonjmp only passes a 32bit int to setjmp, so we store a (64/32bit) exception
pointer in core->exceptionAddr to pass it indirectly.  This was already in place,
but another hack also existed which reserved static room for 2^16 such addresses.

Turns out we only need one.  This patch also fixes MIR to check setjmp's result
for nonzero and load the exception pointer from core->exceptionAddr, like
LIR and the interpreter.