Performance fix for ia32 int->double conversion #474654 (edwsmith+)
As quoted from Moh (Intel) :

"When cvtsi2sd is used for int->double conversion, the destination register is
only partially written. Register renaming cannot properly handle the subsequent
instructions that use the value.  The "partial register stall" problem gets
worse when you have it in a loop, for which case false cross-dependence
relations are created. A solution is to clear out the destination register
before the cvtsi2sd instruction using pxor."