Refactor newfunction and newclass to use CodeWriter.writeOp1() (bug 481244, r=jodyer)
These opcodes have a single operand, the function_id or class_id, and stages also
want to know the function traits or class traits, so it's a sensible move.

Also a quick fix: NullWriter and TeeWriter don't accept NULL, so use a naked CodeWriter
as the stub in the second pass when reporting verify errors.