String fixes (code=daumling, push=stejohns, r=stejohns):
1. UnicodeUtils::Utf8ToUtf16() has an additional strict flag; if set to true,
the method returns -1 on a malformed UTF-8 sequence; if false, the old behavior
remains, where invalid UTF-sequences were stored as characters.
2. String::createUTF8() has this new strict flag, and defaults to true.
3. AvmCore::findInternedString() uses strict == false to be backwards
compatible
4. AvmCore::newStringUtf8() has an additional strict flag, which defaults to
false.
5. FileClass and ByteArrayGlue use AvmCore::newStringUtf8() with strict
explicitly set to false to maintain backwards compatibility.
6. Fix for bug #473995: The headers say that the character buffer must not be
null. There is an assert if the buffer is null. All create methods use a new
static method checkForCachedStrings() that checks if a cached string can be
returned.
7. Fix for bug #473998 along the way: header documentation does not mention
out-of-memory anymore.
8. Changed primitive data in UnicodeUtils to be C99 compatible (except for
uint8* for UTF-8 strings).