More String fixes (code=daumling, review=stejohns, push=stejohns)
1) Rewrote _analyzeUtf8() using UnicodeUtils::Utf8ToUtf16() as a template to
ensure identical behavior.

2) Changed all String constructors to also accept NULL buffers, and to treat
them as zero-length strings; changed header comments.

3) changed checkForCachedStrings() to checkForTinyStrings() and made it a
static String member; returns a zero-length string of the correct width if len
== 0.

4) Fixed bug in createUTF8() that led to a memory overwrite for invalid UTF8
sequences (code did not check for return value of 0 when calling
UnicodeUtils::Utf8ToUcs4).

5) Fixed an old, undiscovered bug in UnicodeUtils::Utf8ToUtf16() that did not
recognize some out-of-range UTF-8 sequences as invalid. Note: this fix may
break existing code! If a ByteArray contained such an out-of-range sequence
(like e.g. FB BF BF BF BF BF for a character value of 0x3FFFF), the sequence
was recognized as a correct sequence, but of a length of 4 bytes.

6) Added test cases to as3/ShellClasses/ByteArray.as to check for invalid,
partial, or out-of-range UTF8 sequences.

The fix for 5) may be problematic. Although it fixes a really obvious bug, the
fix changes the way a ByteArray (or the contents of a file) is converted to
Unicode when assuming UTF-8.