Fix bug 482488 again (r=treilly)
Fixed some bad math when we were recalculating the blockIds and next/prev pointers.
Also modified the logic to only insert a sentinel when we are removing a block from the middle of a contingous section.  In other cases a sentinel will already be present.
In the case that there is already a sentinel on either side, remove one of the extra sentinels.