Adapts the hashtable revisions from TT to reduce size overhead (r=edwsmith)
This removes the vtable from hashtable, splitting it into two classes (InlineHashtable and HeapHashtable). It also compresses internal fields so that InlineHashtable is now 8 bytes on 32-it systems (was 16 bytes formerly). However, it does restrict hashtables to having at most 2^26-1 entries -- unlikely to be a limitation in practice but theoretically achievable in pathological cases.