bug 490900: minor change to runsmokes.py so each test is 1 pass or 1 fail r+brbaker
The smoke tests run tests from the test list (runsmokes.txt) until the 2 min time limit has been exceeded.
We want each test to be 1 pass or 1 fail instead of summing all of the passes/fails in the output.  We want to
be able to track how many tests from the list each platform can run.