PPC back-end branch fix - 489860 (r+edwsmith)
We use JMP(0) to represent a branch to an
unknown destination.  The Assembler patches
this address once the target is resolved, so it must
be able to reach all valid addresses.

On PPC we were using the 24b offset form of the
address, which works most of the time, but
occasionally 32b is needed.  This patch forces
32b form for all branches with unknown destinations.