Improve early binding of constructor (bug 498798 r=rreitmai+)
OP_constructprop contained code to detect when we can safely inline
the fast simple case of constructing a plain object and invoking its
constructor with strongly typed args.  But, OP_construct did not.

This fix factors both opcodes to use a new emitConstruct() which
checks for the necessary conditions and early binds when possible.

The factory pattern implemented with a static method commonly uses
OP_construct and benefits from this change:

class A {
   static function create(...):A {
      return new A()   // now will early bind with this change
   }
}

also note:
v8/typed/raytrace.as does stuff like this in its critical sections.
Flex uses this pattern within the framework, in numerous places.