Remove _startingIns from Assembler (bug 505909 r=rreitmai+)
This field was used in the old error recovery code, to reset _nIns back
to its starting point after an error.  With CodeAlloc, we just return
unused memory back to CodeAlloc, and on the next assembler run, we allocate
new memory.  So there is no need to remember our starting point.