Uses lookup cache to optimize finddef in jit code (bug 515722 r=stejohns+)
This patch enables LookupCache in the jit.  When we emit a call to finddef(), we
allocate a cache slot using the multiname as a key.  At runtime, if we've got a
valid cache entry for that slot, return it and avoid calling the expensive
finddef() lookup.

Allocation of the cache itself is done lazily.  We could do it sooner but for
only little gain (a single compare & branch in the common case) at bigger code
complexity cost.

Cache slots are checked to be valid by comparing a per-slot timestamp with
an avmcore based timestamp that is incremented each time a new ABC is parsed.
(parsing a new abc adds new symbols to the tables searched by finddef, which
can change finddef()'s result).

deferred issue:  some cache slots will be unused if the coresponding finddef
call is found to be dead code later in the compilation pipeline.