Moves callproperty and other instructions to instr module. (bug 517762 r=rreitmai+)
This is just refactoring, removing code duplication, which we'll make use of in
following patches.

The reason for template <class E> is explained in instr.h.  the choices over
what is REALLY_INLINED is to essentially leave the code in the same shape as
before the patch.

More instructions will be moved in subsequent patches, with the goal of consolidating
implementation code in instr[.h,-inlines.h,.cpp].  Inevitably, many trivial operations
will remain in the interpreter and jit, but nontrivial ops should move to this
new central location.