Bug 522558 - NJ merge: absorb some code for handling variadic macros [graydon]
TM explicitly names "no variadic macro support" as a CPP symbol. I think this
approach is a little clearer about what's going on. This patch moves the test
into your platform layer and rewrites the code to use it. There's a
corresponding bit on our side to merge to this state too.