Rename methods in CodegenLIR q->f to reflect nanojit explicit support for float64 (bug 546408 r=stejohns)
Over the past couple months, nanojit has added explicit support for float64 (double)
load/store/immediate instructions, to distinguish them from quad (int64) equivalents.

These renames just reflect the new convention.  Code generation is unchanged.

The rename of fixExceptionsAndLabels() to writeFixExceptionsAndLabels() was an
oversight from when the CodeWriter interface was added.  The implementation in
CodegenLIR is an empty method, so the rename is benign.