Detect branches to the middle of reachable instructions (bug 558442 r=tharwood+ wmaddox+)
The verifier upgrade lost detection of branches to the middles of
instructions.  This restores that check, and loosens it.  Type checking
is performed using a worklist, then blocks are visited in linear order.

We detect an invalid branch as two blocks that overlap.  A branch to
the middle of an unreachable instruction is fine, even almost nonsense,
since if the first 1+ bytes of an instruction are unreachable, the verifier
never even inspects them.

This patch also fixes a latent bug.  If the first bytecode of a function
is an OP_label for a loop, phase two visits the first block twice; the
first time, with the FrameState initialized by the prolog, and the second
time with the FrameState as a result of the back-edge.  This actually results
in code duplication and is not at all what we wanted.

Fixed by creating a block via checkTarget() in phase 1, if the first
instruction is a LIR_label.