Bug 561912 - Verifier's use of offset-from-code_start is too fragile (r=wmaddox+)
Use byte* pc's everywhere except in debug output and exception handler table

In the exception handler table we still need to store int offsets, so we
explicitly capture the byte* pointer to use as a base.  (tryBase).  Everywhere
else we now use byte* pc.

To do this, FrameState->pc (int) was replaced with abc_pc (byte*), CodegenLIR's
blockStates table was modified to use byte* as a key, and the changes in api
were propagated around.