Bug 585433 - Skip deadvars_analyze in non-looping methods.
If the control flow of a method is a DAG, then it is not necessary
to iterate to a fixed point in deadvars(); i.e. we can skip deadvars_analyze(),
and just do the whole analysis with store-removal in deadvars_kill().

This reduces overall verify+jit time by about 3% on a 700MB test set,
which should slightly improve startup time on biggish apps.