Bug 587916 - Cleanup of X87 FP stack code (r=nnethercote+)
1) The "register" FST0 is the sole member of the x87regs register
class. In many places, however, the code is written so as to strongly
suggest that there might be multiple such registers. This patch removes
such conceits, replacing expressions such as (rmask(r) & x87regs)
with (r == FST0), etc.

2) prepareResultReg() has been slightly refactored to make the x87
stack fiddling a bit easier to follow and to remove a fragile assumption.

3) Do not pass the "pop" argument to asm_spill() on non-IA32 platforms.

4) Remove redundant normalization of boolean values.

5) Comment the FPU stack depth consistency check.