Bug 575848 - JIT early binds constructor calls to known builtins (r=stejohns+)
This patch changes the default value of Traits::hasCustomConstruct,
for builtins, to true instead of false.  AS3 native classes that
override ClassClosure::construct() must be declared with

   [native(... customconstruct=true)]

This patch adds support for that in nativegen.py, and updates the
necessary builtings.

A gcc-dependent, debug-only assert will check the C++ _vtbl entry
for ::construct(), and assert if the native class:

 1) overrides ClassClosure::construct() and hasCustomConstruct == false, or
 2) doesn't override ClassClosure::construct() and hasCustomConstruct == true.

Case (1) is a crash; the wrong construct method will be called by JIT code.
To fix, one must add customconstruct=true to the class's [native] decl.

Case (2) is a performance bug; the construct callsite in JIT code will
not be early bound.  One must remove customconstruct=true.