Bug 556357: check input is ASCII before adding modeline (r=fklockii).
I was trying to use this script to do drive-by addition of modelines
to all the actionscript source in test/performance/, but while
sanity-checking the resulting diff, I realized that the code was
blindly adding content to the beginning of files with a byte-order
mark (BOM).  That would be bonkers, and so I am trying to safeguard
against such events by first using the 'file' command to double check
that the input file format is actually ASCII before mucking with it.

Its beyond the scope of a script like this to update files with a BOM.
They can just be updated by hand, whenever one wants to perform such
updating.

(My main goal in adding the modelines is to guard against accidental
re-introduction of tab characters to the files.)