Bug 715105: Fix for AMF changes inject bug because Domain Env not set properly (r=fklockii) . ATS10AS3 broke after patches for Bug 673355 was integrated to player mainline.
(We didn't catch this because the tests that failed are not part of the set that are run automatically.)

Kapil has a fix in his player sandbox.

This bug is to stash conversation related to this bug and the patches for it, rather than tacking it onto Bug 673355 which already has over 75 comments.