Bug 729336: add manifest.mk for lzma.
This manifest.mk appears to have been omitted from msong's
original submissions.

Also, as an addendum to the previous commit (CL 1037814), in
addition to the integration and resolve commands I described
there, I also put in some extra modifications to get the code
building on xplat.  (I am not used to Perforce, where "p4
change" is not the same thing as a local "hg commit" or even a
local MQ "hg qnew"; subsequent changes to the local files end
up as part of the final submission, for better or for worse.)