Bug 732503: Remove FRR builds from buildbot (r=trbaker).
This changeset was cherry-picked from a Perfarce clone of the FR-Avm
branch in P4, via the following command (issued with the knowledge
that the desired changeset was revision 60 aka CL 1039347):

( cd ~/Dev/FlashPFarce/fklockii-MacPro-FRAvmJustTR && \
  hg export -r 60                                     \
) | hg import -

Of course that just yields the original changeset message (the first
line above).  I have manually hg-qimport'ed and hg-qrefresh'ed the
changeset since then to include this documentation.