Bug 702552, 740903: VM stack pointer metrics and overflow recovery hooks (r=edwsmith)
The patch addresses the handling of the StackOverflowError in AS3, and in native code
to the extent that we are relying on the AS3 checks.

This is principally a fix for WE 2949577, with changes to the Flash Player avmglue code.
The avmplus enhancements consist of hooks to support better recovery by the host,
and instrumentation that is useful both for testing and for tuning of allocation margins.
On the player side, the patch was reviewed by Jeff Mott.  Edwin reviewed only the Tamarin changes.

CL@1048913