Summary: Fix a performance problem with some Alchemy generated SWFs Severity: ignore Detailed Description: This is a slightly tweaked version of a fix that Edwin Smith came up with a year ago, in his words:
During phase 1, the verifier iterates to a fixed point while checking the code, using a worklist algorithm.  The worklist is not sorted, so it is possible for a block to be revisited far more often than necessary.

If the ABC basic blocks are roughly in reverse postorder, then this problem is worst when the blocks in the worklist are in high-to-low abc order.

It turns out the main cause of re-queues is int-typed variables
changing from notNull=true -> false.  Really, the notnull bit doesn't
apply to integers, so we need to force them =true for numbers or
just ignore the bool.

this patch makes them true by default but doesn't force them.  I think
there still is noise with uint, but the requeues are way down

This visit order assumes the ABC is approximately ideal (nearly reverse postorder) because I didn't want to add the complexity of really finding the CFG and really visiting in reverse postorder.  However, if we did that, the asymptotic time would probably go down even more.


Dev/QA impact: the usual
QA testing notes: N/A
API change:N/A
Bugs fixed: N/A
Doc impact: N/A
Smokes passed: avmshell acceptance + build smokes + Alchemy smokes
Reviewer: Stan Switzer

CL@1126621